Refactoring and cleanup around transaction appenders
* Collapses the AbstractPhysicalTransactionAppender hierarchy.
* Removes non-batched appending and obsoletes the `batched_writes`
* configuration option.
* Rename a couple of methods for clarity.
* Move InMemoryLogChannel to test sources.
