+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/log/TransactionAppender.java
@@ -62,7 +62,7 @@ public interface TransactionAppender
      *
      * @param transaction transaction representation to append.
      * @param transactionId the appended transaction is expected to have.
-     * @return {@link Commitment} that should be marked as {@link Commitment#transactionCommitted() committed}
+     * @return {@link Commitment} that should be marked as {@link Commitment#publishAsCommitted() committed}
      * after a point where the log has been manually {@link #force() forced}.
      * @throws IOException if there was a problem appending the transaction, or if the transaction id
      * generated by this appender doesn't match.
